code cleanup v3
