--- src/RefDocGen/TemplateGenerators/Default/TemplateModelCreators/ParameterTMCreator.cs
@@ -1,59 +1,0 @@
-using RefDocGen.CodeElements.Abstract.Members;
-using RefDocGen.TemplateGenerators.Default.TemplateModels.Members;
-using RefDocGen.TemplateGenerators.Tools;
-using RefDocGen.TemplateGenerators.Tools.DocComments.Html;
-using RefDocGen.TemplateGenerators.Tools.Keywords;
-using RefDocGen.TemplateGenerators.Tools.TypeName;
-
-namespace RefDocGen.TemplateGenerators.Default.TemplateModelCreators;
-
-/// <summary>
-/// Class responsible for creating template models representing the individual method parameters.
-/// </summary>
-internal class ParameterTMCreator : BaseTMCreator
-{
-    public ParameterTMCreator(IDocCommentTransformer docCommentTransformer) : base(docCommentTransformer)
-    {
-    }
-
-    /// <summary>
-    /// Creates a <see cref="ParameterTM"/> instance based on the provided <see cref="IParameterData"/> object.
-    /// </summary>
-    /// <param name="parameterData">The <see cref="IParameterData"/> instance representing the parameter.</param>
-    /// <returns>A <see cref="ParameterTM"/> instance based on the provided <paramref name="parameterData"/>.</returns>
-    internal ParameterTM GetFrom(IParameterData parameterData)
-    {
-        List<Keyword> modifiers = [];
-
-        if (parameterData.IsInput)
-        {
-            modifiers.Add(Keyword.In);
-        }
-
-        if (parameterData.IsOutput)
-        {
-            modifiers.Add(Keyword.Out);
-        }
-
-        if (RefKeyword.IsPresentIn(parameterData))
-        {
-            modifiers.Add(Keyword.Ref);
-        }
-
-        if (parameterData.IsParamsCollection)
-        {
-            modifiers.Add(Keyword.Params);
-        }
-
-        string? defaultValue = parameterData.DefaultValue == DBNull.Value
-            ? null
-            : LiteralValueFormatter.Format(parameterData.DefaultValue);
-
-        return new ParameterTM(
-            parameterData.Name,
-            CSharpTypeName.Of(parameterData.Type),
-            ToHtmlString(parameterData.DocComment),
-            modifiers.GetStrings(),
-            defaultValue);
-    }
-}
